Beta 1 - And Possibly Final Version
I don't really remember where it was last time, but it now
plots the data for multiple files and automatically calculates the
minimum value within the standard range. (+5mm - End) It allows
you to click twice to specify a custom range and will update
the table of values. You may then click the copy button and all
visible data will be copied to the clipboard, sutable to be pasted
into a spreadsheet or text document. There is now an option to
erase the current data when new data files are opened.